Remoção de classes não mais utilizadas
