--- app/src/main/res/layout/activity_aluno.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/fundo_night"
-    tools:context="online.padev.kariti.AlunoActivity">
-
-    <include
-        android:id="@+id/include"
-        layout="@layout/appbarquestion"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="0dp" />
-
-    <Button
-        android:id="@+id/buttonCadAluno"
-        style="@style/BUTTONSTYLE"
-        android:layout_width="320dp"
-        android:layout_height="wrap_content"
-        android:background="@drawable/button_selector"
-        android:padding="5dp"
-        android:text="@string/buttonCadAluno"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.494"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/divider4"
-        app:layout_constraintVertical_bias="0.307" />
-
-    <Button
-        android:id="@+id/buttonVisualAluno"
-        style="@style/BUTTONSTYLE"
-        android:layout_width="320dp"
-        android:layout_height="wrap_content"
-        android:background="@drawable/button_selector"
-        android:padding="5dp"
-        android:text="@string/buttonVisuAluno"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.494"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/buttonCadAluno"
-        app:layout_constraintVertical_bias="0.153" />
-
-    <ImageView
-        android:id="@+id/imageView8"
-        style="@style/COLORIMGICONE"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/imgBtnSeta"
-        app:layout_constraintBottom_toBottomOf="@+id/buttonCadAluno"
-        app:layout_constraintEnd_toEndOf="@+id/buttonCadAluno"
-        app:layout_constraintHorizontal_bias="0.958"
-        app:layout_constraintStart_toStartOf="@+id/buttonCadAluno"
-        app:layout_constraintTop_toTopOf="@+id/buttonCadAluno"
-        app:layout_constraintVertical_bias="0.352"
-        app:srcCompat="@mipmap/ic_arrownext" />
-
-    <ImageView
-        android:id="@+id/imageView12"
-        style="@style/COLORIMGICONE"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:contentDescription="@string/textViewAluno"
-        app:layout_constraintBottom_toBottomOf="@+id/buttonCadAluno"
-        app:layout_constraintEnd_toEndOf="@+id/buttonCadAluno"
-        app:layout_constraintHorizontal_bias="0.062"
-        app:layout_constraintStart_toStartOf="@+id/buttonCadAluno"
-        app:layout_constraintTop_toTopOf="@+id/buttonCadAluno"
-        app:layout_constraintVertical_bias="0.357"
-        app:srcCompat="@mipmap/student" />
-
-    <View
-        android:id="@+id/divider4"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="?android:attr/listDivider"
-        app:layout_constraintTop_toBottomOf="@+id/include"
-        tools:ignore="MissingConstraints"
-        tools:layout_editor_absoluteX="2dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
