--- app/src/main/res/layout/activity_camerax.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/fundo_night"
-    tools:context="online.padev.kariti.CameraxActivity">
-
-    <TextView
-        android:id="@+id/txtViewinfoCameraX"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="30dp"
-        android:text="@string/informacaoCamera"
-        android:textSize="20sp" />
-
-    <androidx.camera.view.PreviewView
-        android:id="@+id/previewCameraX"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/txtViewinfoCameraX"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentEnd="true"
-        android:layout_margin="10dp">
-
-    </androidx.camera.view.PreviewView>
-
-    <ImageButton
-        android:contentDescription="@string/iconeCaptura"
-        android:id="@+id/buttonCaptureX"
-        android:layout_width="81dp"
-        android:layout_height="65dp"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="30sp"
-        android:scaleType="centerCrop"
-        android:src="@drawable/baseline_camera_24"
-        style="@style/BUTTONSTYLEdark"
-        android:background="?attr/selectableItemBackground"
-        app:tint="@color/white" />
-
-    <ImageButton
-        android:contentDescription="@string/iconeFlash"
-        android:id="@+id/buttonFlashX"
-        android:layout_width="48sp"
-        android:layout_height="48sp"
-        android:layout_below="@+id/txtViewinfoCameraX"
-        android:layout_alignParentEnd="true"
-        android:layout_margin="20sp"
-        android:src="@drawable/baseline_bolt_24"
-        android:background="?attr/selectableItemBackground"
-        app:tint="@color/white"/>
-    
-    <ImageButton
-        android:contentDescription="@string/iconeAround"
-        android:id="@+id/aroundCameraX"
-        android:layout_width="48sp"
-        android:layout_height="48sp"
-        android:layout_below="@id/buttonFlashX"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="20sp"
-        android:src="@drawable/baseline_cameraswitch_24"
-        android:background="?attr/selectableItemBackground"
-        app:tint="@color/white"/>
-
-</RelativeLayout>
